registration/requirements.txt
Django==1.9.2
django-localflavor==1.2
wheel==0.24.0
